Working code (BUT)
This code is running correctly however, the actual warnings are not
being shown in the database. Also, parallel processing is not accounted
for yet. 'IF NOT CREATED' constraint has been added to the sql code,
this means the database does not have to be deleted before each run.
However, as a seed is set, the records within the table will have to be
deleted otherwise the unique constraint is violated.